% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpSetColor.R
\name{UpSetColor}
\alias{UpSetColor}
\title{UpSetColor - Visualization of intersecting sets with colored highlights}
\usage{
UpSetColor(
  data,
  mode = c("union", "intersect", "distinct")[3],
  result = c("plot", "list")[1],
  set_order = c("decreasing", "increasing", "as.given", "as.given.reverse")[2],
  comb_order = NULL,
  min_set_size = NULL,
  top_n_sets = NULL,
  min_comb_degree = NULL,
  max_comb_degree = NULL,
  min_comb_size = NULL,
  max_comb_size = NULL,
  top_comb = NULL,
  color.1 = "black",
  color.0 = "#cfcfcf",
  comb_highlight = NULL,
  color.highlight = "red",
  numbers.size = NULL,
  size.line = 0.5,
  size.dot = 3,
  color.bar.sets = "black",
  color.bar.comb = "black",
  plot.title.size = 14,
  size.comb.axis.y = 8,
  size.matrix.axis.y = 8,
  size.Set.axis.x = 8,
  title.plot = "Upset Plot",
  title.comb.axis.y = "Number of elements",
  title.Set.axis.x = "Set size",
  title.matrix.y = "Sets",
  scale.Set.y.log10 = FALSE,
  title.hjust = 0.5,
  heights = c(0.35, 0.65),
  widths = c(0.7, 0.3),
  color.stripes = c("#FFFFFF", "#F0F0F0"),
  verbose = FALSE
)
}
\arguments{
\item{data}{A list of sets (set of the list of sets is a vector), or a binary matrix or data frame (where rows are elements and columns are sets).}

\item{mode}{The mode used to calculate the set combinations. Possible options are "distinct" (default), "intersect", and "union". In distinct mode, 1 means in that set and 0 means not in that set; in intersect mode: 1 means in that set and 0 is not taken into account; in union mode, as in the intersect mode, 1 means in that set and 0 is not taken into account. When there are multiple 1 in union mode, the relationship is OR. Under the union mode, the seven combination sets can overlap.}

\item{result}{Possible options are "plot" (default), "list.plot" and "list.data". With "list.plot", a list with the 3 ggplot objects that compose the UpSet plot is returned. With "list.data", a list with the data used as input of the plots is returned.}

\item{set_order}{Possible options are "increasing" (default), "decreasing", "as.given" and "as.given.reverse". It controls the order of the sets in the Set bar plot (bottom right) and the matrix plot (bottom left). "increasing" and "decreasing" sort the bars by their size, "as.given" produces no sorting. "as.given.reverse" reverses the set order.}

\item{comb_order}{Vector of integers to customize manually the order of the combination bar plot (top).}

\item{min_set_size}{The minimum number of elements needed in a set. If a set does not have more than this number, it will not be included in the UpSet plot. If NULL, no set filtering is performed according to this argument.}

\item{top_n_sets}{Integer denoting that only the n-th largest sets will be used in the UpSet plot. If NULL, no set filtering is performed according to this argument.}

\item{min_comb_degree}{The minimum number of sets compared. Set combinations involving less than this number of sets will not be shown in the UpSet plot.}

\item{max_comb_degree}{The maximum number of sets compared. Set combinations involving more than this number of sets will not be shown in the UpSet plot.}

\item{min_comb_size}{The minimum number of elements in the set combination. Set combinations below this number will not be shown in the UpSet plot.}

\item{max_comb_size}{The maximum number of elements in the set combination. Set combinations above this number will not be shown in the UpSet plot.}

\item{top_comb}{The number of element combinations to display in the combination bar plot (top) and the matrix plot (bottom left). The default is 20. To display all combinations, this argument must be set at 2^length(data)-1 (if data is a list) or 2^ncol(data)-1 (if data is a binary matrix or data frame).}

\item{color.1}{String with the colors used to color the dots involved in every set combination. It must be of the same size as the number of displayed combinations in the upper plot, or be of length 1. If NULL, dots will be colored in black.}

\item{color.0}{String with the color used to color the dots not involved in the set combinations. It is of length 1. Default is "#cfcfcf".}

\item{comb_highlight}{The bar positions in the combination bar plot (top) that will be highlighted. The involved Sets will also be highlighted in the other plots. If NULL, none of the set combinations is highlighted.}

\item{color.highlight}{Optional. Color used to highlight the set combinations. The default is red.}

\item{numbers.size}{The font size for the bar labels in the combination plot (top). If NULL (default), the bar labels will not be displayed.}

\item{size.line}{The width of the line connecting the "1" dots in the matrix plot. The default is 0.5.}

\item{size.dot}{The dot size in the matrix plot. The default is 3.}

\item{color.bar.sets}{String with the colors used to color the bars in the bottom-right plot. It must be of the same size as the total number of sets, or be of length 1. If NULL, the bars will be colored in black.}

\item{color.bar.comb}{String with the colors used to color the bars in the top plot. It must be of the same size than the total number of displayed combinations, or be of length 1. If NULL, the bars will be colored in black.}

\item{plot.title.size}{Title font size. Default is 14.}

\item{size.comb.axis.y}{Y-axis title font size in the top plot. Default is 8.}

\item{size.matrix.axis.y}{Y-axis title font size in the matrix plot. Default is 8.}

\item{size.Set.axis.x}{X-axis title font size in the bottom-right plot. Default is 8.}

\item{title.plot}{String with the plot title. Default is "UpSet plot".}

\item{title.comb.axis.y}{String with the title used in the y-axis of the top plot. The default is "Number of elements".}

\item{title.Set.axis.x}{String with the title used in the y-axis of the bottom-right plot. The default is "Set size".}

\item{title.matrix.y}{String with the title used in the y-axis of the matrix plot. The default is "Sets".}

\item{scale.Set.y.log10}{If FALSE, the y-axis of the top plot is decimal. If TRUE, the scale is log10.}

\item{title.hjust}{Parameter for horizontal adjustment of the main title. The default is 0.5.}

\item{heights}{Plot heights in the layout. The default is c(0.35, 0.65).}

\item{widths}{Plot widths in the layout. The default is c(0.7,0.3).}

\item{color.stripes}{The two colors used to fill the rectangles in the background. The default is c("#FFFFFF", "#F0F0F0")}

\item{verbose}{If TRUE, a progress bar and the spent time are shown.}
}
\value{
A plot (if results = 'plot'), a list of ggplot objects (if results = 'list.plot') or a list of data.frames (if result = 'list.data').
}
\description{
Make the UpSet plot.
}
\examples{
set.seed(123)
lt = list(a = sample(letters, 5),
         b = sample(letters, 10),
         c = sample(letters, 15))
UpSetColor(data = lt, top_comb=7)
}
